requirements.txt
fastapi
uvicorn
python-multipart
motor
pytesseract
pillow
